+++ src/modules/renders/render-auth.js
@@ -1,0 +1,10 @@
+import auth from '../requests/auth.js';
+
+/**
+ * implementation rendering of authorization
+ * @param {htmlElement} parent - parent element
+ * @param {json} config - configuration
+ */
+export default (parent, config) => {
+    auth(parent, config);
+};
