+++ src/modules/renders/render-chat.js
@@ -1,0 +1,23 @@
+import chat from '../../templates/chat.js';
+import getParentElement from '../getParentElement.js';
+import logout from '../requests/logout.js';
+
+/**
+ * implementation rendering of main page
+ * @param {htmlElement} parent - parent element
+ * @param {json} config - configuration
+ * @param {int} userId - User identifier
+ */
+export default (parent, config, userId) => {
+    getParentElement().innerHTML = chat();
+
+    if (userId) {
+        document.querySelector('.header__user-photo').src = './assets/img/geva.png';
+    }
+
+    document.querySelector('.logout').addEventListener('click', (e) => {
+        e.preventDefault();
+
+        logout(parent, config);
+    });
+};
