+++ src/modules/requests/config.js
@@ -1,0 +1,7 @@
+export default {
+    auth: '/auth/',
+    login: '/login/',
+    signup: '/signup/',
+    chat: '/users/',
+    logout: '/logout/',
+};
